+++ apps/meteor/client/components/message/content/attachments/file/GenericFileAttachment.tsx
@@ -78,12 +78,7 @@ const GenericFileAttachment = ({
 					<MessageGenericPreviewContent
 						thumb={<MessageGenericPreviewIcon name='attachment-file' type={format || getFileExtension(title)} />}
 					>
-						<MessageGenericPreviewTitle
-							download={!!openDocumentViewer}
-							externalUrl={getExternalUrl()}
-							onClick={handleTitleClick}
-							data-qa-type='attachment-title-link'
-						>
+						<MessageGenericPreviewTitle download={!!openDocumentViewer} externalUrl={getExternalUrl()} onClick={handleTitleClick}>
 							{title}
 						</MessageGenericPreviewTitle>
 						{size && (