+++ apps/meteor/client/views/room/contextualBar/RoomFiles/hooks/useFilesList.ts
@@ -41,6 +41,7 @@ export const useFilesList = ({ rid, type, text }: { rid: Required<IUpload>['rid'
 				...file,
 				uploadedAt: file.uploadedAt ? new Date(file.uploadedAt) : undefined,
 				modifiedAt: file.modifiedAt ? new Date(file.modifiedAt) : undefined,
+				expiresAt: file.expiresAt ? new Date(file.expiresAt) : undefined,
 			}));
 
 			for await (const file of items) {
@@ -55,7 +56,7 @@ export const useFilesList = ({ rid, type, text }: { rid: Required<IUpload>['rid'
 								type: decrypted.type,
 							}),
 						);
-						decrypted.path = `/file-decrypt${decrypted.path}?key=${key}`;
+						decrypted.path = `/file-decrypt${decrypted.path}?key=${encodeURIComponent(key)}`;
 						Object.assign(file, decrypted);
 					}
 				}