+++ apps/uikit-playground/src/Context/action/duplicateScreenAction.ts
@@ -1,14 +1,11 @@
-import { ActionTypes } from "../reducer";
+import { ActionTypes } from '../reducer';
 
 export type DuplicateScreenAction = {
-  type: ActionTypes.DuplicateScreen,
-  payload: { id: string, name?: string },
+	type: ActionTypes.DuplicateScreen;
+	payload: { id: string; name?: string };
 };
 
-export const duplicateScreenAction = (payload: {
-  id: string,
-  name?: string,
-}): DuplicateScreenAction => ({
-  type: ActionTypes.DuplicateScreen,
-  payload,
+export const duplicateScreenAction = (payload: { id: string; name?: string }): DuplicateScreenAction => ({
+	type: ActionTypes.DuplicateScreen,
+	payload,
 });