+++ packages/fuselage-ui-kit/src/elements/ChannelsSelectElement/MultiChannelsSelectElement.tsx
@@ -20,7 +20,7 @@ const MultiChannelsSelectElement = ({ block, context }: MultiChannelsSelectProps
 
 	const handleChange = useCallback(
 		(value: string | string[]) => {
-			if (Array.isArray(value)) action({ target: { value } });
+			if (Array.isArray(value)) void action({ target: { value } });
 		},
 		[action],
 	);