+++ packages/message-parser/src/definitions.ts
@@ -83,6 +83,18 @@ export type Quote = {
 	value: Paragraph[];
 };
 
+export type Spoiler = {
+	type: 'SPOILER';
+	value: Array<
+		Link | Emoji | UserMention | ChannelMention | InlineCode | Italic | Bold | Strike | Timestamp | Color | Image | InlineKaTeX | Plain
+	>;
+};
+
+export type SpoilerBlock = {
+	type: 'SPOILER_BLOCK';
+	value: Paragraph[];
+};
+
 export type Markup = Italic | Strike | Bold | Plain | ChannelMention;
 export type MarkupExcluding<T extends Markup> = Exclude<Markup, T>;
 
@@ -163,6 +175,7 @@ export type Timestamp = {
 
 export type Types = {
 	BOLD: Bold;
+	SPOILER: Spoiler;
 	PARAGRAPH: Paragraph;
 	PLAIN_TEXT: Plain;
 	ITALIC: Italic;
@@ -185,11 +198,13 @@ export type Types = {
 	LIST_ITEM: ListItem;
 	IMAGE: Image;
 	LINE_BREAK: LineBreak;
+	SPOILER_BLOCK: SpoilerBlock;
 };
 
 export type ASTNode =
 	| BigEmoji
 	| Bold
+	| Spoiler
 	| Paragraph
 	| Plain
 	| Italic
@@ -199,6 +214,7 @@ export type ASTNode =
 	| InlineCode
 	| Heading
 	| Quote
+	| SpoilerBlock
 	| Link
 	| UserMention
 	| ChannelMention
@@ -211,6 +227,7 @@ export type TypesKeys = keyof Types;
 export type Inlines =
 	| Timestamp
 	| Bold
+	| Spoiler
 	| Plain
 	| Italic
 	| Strike
@@ -223,6 +240,6 @@ export type Inlines =
 	| Color
 	| InlineKaTeX;
 
-export type Blocks = Code | Heading | Quote | ListItem | Tasks | OrderedList | UnorderedList | LineBreak | KaTeX;
+export type Blocks = Code | Heading | Quote | SpoilerBlock | ListItem | Tasks | OrderedList | UnorderedList | LineBreak | KaTeX;
 
 export type Root = Array<Paragraph | Blocks> | [BigEmoji];